added new de1 project files. updated fringe so that it supports the 32-bit system on de1.

+++ resources/synth/de1.sw-resources/FringeContextBase.h
@@ -11,25 +11,25 @@ public:
 
     virtual void load() = 0;
 
-    virtual uint64_t malloc(size_t bytes) = 0;
+    virtual uint32_t malloc(size_t bytes) = 0;
 
-    virtual void free(uint64_t buf) = 0;
+    virtual void free(uint32_t buf) = 0;
 
-    virtual void memcpy(uint64_t devmem, void *hostmem, size_t size) = 0;
+    virtual void memcpy(uint32_t devmem, void *hostmem, size_t size) = 0;
 
-    virtual void memcpy(void *hostmem, uint64_t devmem, size_t size) = 0;
+    virtual void memcpy(void *hostmem, uint32_t devmem, size_t size) = 0;
 
     virtual void run() = 0;
 
-    virtual void writeReg(uint32_t reg, uint64_t data) = 0;
+    virtual void writeReg(uint32_t reg, uint32_t data) = 0;
 
-    virtual uint64_t readReg(uint32_t reg) = 0;
+    virtual uint32_t readReg(uint32_t reg) = 0;
 
-    virtual uint64_t getArg(uint32_t arg, bool isIO) = 0;
+    virtual uint32_t getArg(uint32_t arg, bool isIO) = 0;
 
-    virtual uint64_t getArg64(uint32_t arg, bool isIO) = 0;
+    virtual uint32_t getArg64(uint32_t arg, bool isIO) = 0;
 
-    virtual void setArg(uint32_t reg, uint64_t data, bool isIO) = 0;
+    virtual void setArg(uint32_t reg, uint32_t data, bool isIO) = 0;
 
     virtual void setNumArgIns(uint32_t number) = 0;
